userLogs
Underpinning for our userLogs page. Source for reference is linked
above. This is using data from a JSON file to fill a flatList with first
and last names.

These names are coming from a API:
https://randomuser.me/api?results=10

I encourage you to take a look at the generated JSON file to understand
it. We can use this approach for listing Query results from the
NutritionX API.